% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provParse.R
\name{prov.parse}
\alias{prov.parse}
\title{Parses the information in a prov.json file and makes it available to be 
when accessed through the access functions.}
\usage{
prov.parse(prov.input, isFile = T)
}
\arguments{
\item{prov.input}{A path to a json file that has been generated by RDataTracker
or a string that is in prov.json format}

\item{isFile}{A logical that determines whether or not the input has to be read 
in from the file system}
}
\value{
Does not return a value, but opens up the use of the access functions
}
\description{
Parses the information in a prov.json file and makes it available to be 
when accessed through the access functions.
}
\examples{
prov.parse(system.file ("testdata", "ddg.json", package="provParseR", mustWork=TRUE))
}
