% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provParse.R
\docType{class}
\name{ProvInfo-class}
\alias{ProvInfo-class}
\alias{ProvInfo}
\title{Collection of information gathered from parsing a PROV file}
\description{
This is the class that stores provenance information.  It is created by 
prov.parse.  Rather than access the slots directly, it is better to use
the access functions the package provides.
}
\section{Slots}{

\describe{
\item{\code{proc.nodes}}{the procedure nodes}

\item{\code{data.nodes}}{the data nodes}

\item{\code{func.nodes}}{the function nodes}

\item{\code{proc.proc.edges}}{control flow edges}

\item{\code{proc.data.edges}}{output data edges}

\item{\code{data.proc.edges}}{input data edges}

\item{\code{func.proc.edges}}{function use edges}

\item{\code{func.lib.edges}}{function library edges}

\item{\code{agents}}{tool that created the provenance}

\item{\code{args}}{arguments passed when provenance was created}

\item{\code{envi}}{environmental information}

\item{\code{libs}}{libraries}

\item{\code{scripts}}{scripts executed}
}}

\seealso{
The parse function, which creates the ProvInfo object, \code{\link{prov.parse}}

The access functions, including \code{\link{get.environment}}
}
