% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provParse.R
\name{access}
\alias{access}
\alias{get.environment}
\alias{get.libs}
\alias{get.scripts}
\alias{get.proc.nodes}
\alias{get.data.nodes}
\alias{get.func.nodes}
\alias{get.proc.proc}
\alias{get.data.proc}
\alias{get.proc.data}
\alias{get.func.proc}
\alias{get.func.lib}
\alias{get.all.df}
\title{Access functions that will retrieve information that has been parsed from the prov.json file.
The function prov.parse MUST be run before using these functions.}
\usage{
get.environment()

get.libs()

get.scripts()

get.proc.nodes()

get.data.nodes()

get.func.nodes()

get.proc.proc()

get.data.proc()

get.proc.data()

get.func.proc()

get.func.lib()

get.all.df()
}
\value{
The data frame of the requested information.
}
\description{
Access functions that will retrieve information that has been parsed from the prov.json file.
The function prov.parse MUST be run before using these functions.
}
\examples{
prov.parse(system.file ("testdata", "ddg.json", package="provParseR", mustWork=TRUE))
get.proc.nodes()
get.data.nodes()
get.func.nodes()
get.proc.proc()
get.data.proc()
get.proc.data()
get.func.proc()
get.func.lib()
get.libs()
get.scripts()
get.environment()
get.all.df()

}
