% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provParse.R
\name{prov.parse}
\alias{prov.parse}
\title{Provenance parser}
\usage{
prov.parse(prov.input, isFile = T)
}
\arguments{
\item{prov.input}{A path to a json file that has been generated by RDataTracker
or provR or a string that is in prov.json format.}

\item{isFile}{A logical that determines whether or not the input prov.input parameter
contains a file name or contains the provenance in a string.}
}
\value{
Does not return a value.  Accessor functions can be called to get 
   the desired portions of the provenance
}
\description{
Parses the provenance stored in a prov.json file.  These files are created
by the provR and RDataTracker packages.  The provParseR package also defines
a number of functions that extract and return information from the parsed
provenance.
}
\examples{
prov.parse(system.file ("testdata", "prov.json", package="provParseR", mustWork=TRUE))
}
\seealso{
The access functions, including \code{\link{get.environment}}
}
