% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provParse.R
\name{prov.parse}
\alias{prov.parse}
\title{Parses the information in a prov.json file and makes it available to be 
when accessed through the access functions.}
\usage{
prov.parse(filename)
}
\arguments{
\item{filename}{A path to a json file that has been generated by RDataTracker}
}
\value{
Does not return a value, but opens up the use of the access functions
}
\description{
Parses the information in a prov.json file and makes it available to be 
when accessed through the access functions.
}
\examples{
prov.parse(system.file ("testdata", "ddg.json", package="provParseR", mustWork=TRUE))
}
